optmize profile query

+++ app/http/controllers/ProfileController.py
@@ -5,19 +5,11 @@ from app.User import User
 from app.Follow import Follow
 
 class ProfileController:
-    """ProfileController
-    """
+    """ProfileController."""
 
     def show(self, request: Request):
         user = User.where('username', request.param('username')).first()
-        follow = Follow.where('user_id', user.id).where('follower_id', request.user().id).first()
-        profile = { 
-            'username': user.username,
-            'image': user.image,
-            'bio': user.bio,
-            'following': True if follow else False
-        }
-        return {'profile': profile}
+        return {'profile': user.payload(request.user())}
 
     def follow(self, request: Request):
         user = User.where('username', request.param('username')).first()
@@ -25,22 +17,11 @@ class ProfileController:
             user_id=user.id,
             follower_id=request.user().id
         )
-        profile = {
-            'username': user.username,
-            'image': user.image,
-            'bio': user.bio,
-            'following': True if follow else False
-        }
-        return {'profile': profile}
+        return {'profile': user.payload(request.user(), follow)}
     
     def unfollow(self, request: Request):
         user = User.where('username', request.param('username')).first()
         follow = Follow.where('user_id', user.id).where('follower_id', request.user().id).first()
-        follow.delete()
-        profile = {
-            'username': user.username,
-            'image': user.image,
-            'bio': user.bio,
-            'following': False
-        }
-        return {'profile': profile}
+        if follow:
+            follow.delete()
+        return {'profile': user.payload(request.user())}
